reorganize constants (internal & public)
