+++ src/constants-internal.ts
@@ -1,6 +1,5 @@
-/******************************************************************************
- * VERSIONS
- */
+/** @internalRemarks Source: @carto/constants, @deck.gl/carto */
+export const DEFAULT_API_BASE_URL = 'https://gcp-us-east1.api.carto.com';
 
 /**
  * Current version of @carto/api-client.
@@ -10,6 +9,18 @@ export const API_CLIENT_VERSION = __CARTO_API_CLIENT_VERSION;
 
 /** @internal */
 export const V3_MINOR_VERSION = '3.4';
+
+/** @internalRemarks Source: @carto/constants, @deck.gl/carto */
+export const DEFAULT_GEO_COLUMN = 'geom';
+
+/**
+ * Fastly default limit is 8192; leave some padding.
+ * @internalRemarks Source: @deck.gl/carto
+ */
+export const DEFAULT_MAX_LENGTH_URL = 7000;
+
+/** @internalRemarks Source: @deck.gl/carto */
+export const DEFAULT_TILE_RESOLUTION = 0.5;
 
 /**
  * @internalRemarks Source: @deck.gl/carto
